scraper/requirements.txt
pymongo>=3.11.4 
undetected-chromedriver>=3.0.2 
beautifulsoup4>=4.9.3 
python-dotenv>=0.18.0 
dnspython>=2.1.0
lxml>=4.6.3
